Big file restructure. All profile components and the reduxProfile have been merged into one directory. Profile/Profile.js serves it up, all components are in subdirectories. The main profile folder also has old-Github.js & .css, which I am mining for parts to integerate GitHub auth on profile page. UserAction folder also has an old.js for similar purposes.
