+++ client/src/components/Profile/old-GitHub.css
@@ -1,0 +1,29 @@
+/* OLD GITHUB PROFILE PAGE */
+.profContainer {
+    width: 100vw;
+    display: flex;
+    flex-direction: column;
+}
+  
+.profCard {
+    margin: 0.75rem 0;
+    background-color: rgb(245, 245, 248);
+    padding: 25px 20px;
+    border-radius: 4px;
+    box-shadow: 1px 1px 3px rgba(0, 0, 0, 0.1);
+    border: 1px solid rgba(13, 37, 145, 0.1);
+}
+
+.cardContainer {
+  display: flex;
+  flex-direction: column;
+  margin: 1rem 10% 5rem 10%;
+}
+
+/* MEDIA QUERIES aren't working yet */ 
+/* Screen > Portrait Tablet */
+@media (min-width: 960px) {
+    .cardContainer {
+      margin: 1rem auto 5rem auto;
+    }
+}
