+++ client/src/components/Profile/old-GitHub.js
@@ -1,11 +1,16 @@
-// PROFILE PAGE
+//////////////////////////////////////////////////////
+// OLD GITHUB AUTH'D PROFILE PAGE
+// Not part of production, might use parts,
+// Particularly the OldUserAction in UserAction/old.js
+//////////////////////////////////////////////////////
+
 import React from "react";
 // import React, { createContext, useReducer} from "react";
 import NavBar from "../NavBar";
 import UserCard from "./UserCard";
 import SocialCard from "./SocialCard";
 import AboutCard from "./AboutCard";
-import UserActions from "./UserActions";
+import OldUserAction from "./UserAction/old.js";
 // import PostsCard from "./PostsCard";
 import "./style.css";
 
@@ -17,9 +22,8 @@ const Profile = () => {
       <div className="cardContainer">
         <UserCard />
         <SocialCard />
-        {/* To Do <PostsCard /> */}
         <AboutCard />
-        <UserActions />
+        <OldUserAction />
       </div>
     </div>
   );
