+++ client/src/components/Profile/style.css
@@ -1,8 +1,42 @@
-/* PROFILE PAGE */
-.profContainer {
-    width: 100vw;
-    display: flex;
-    flex-direction: column;
+/* PROFILE - Main */ 
+/* BUT this isn't even linked into Profile.js right now,
+    but it still somehow affects the page???
+    the css iss an absolute mess : ( */ 
+#social {
+  flex-direction: row;
+}
+
+.socialBtn {
+  transition: 0.3s;
+  background-color: inherit;
+  border: none;
+}
+
+.socialBtn i {
+  color: #384047;
+  font-size: 3rem;
+
+}
+.socialBtn:hover, .socialBtn:focus {
+  letter-spacing: 0.3rem;
+  box-shadow: 3px 5px 3px 0 rgba(0, 0, 0, 1);
+  background-color: inherit;
+}
+
+#date{
+  float: right;
+  margin: 4rem auto -3.5rem;
+  font-family: inherit;
+  font-size: 0.75rem;
+  color: #545454;
+}
+
+.card {
+  box-shadow: 0 8px 12px 0 rgba(10, 35, 90, 0.2);
+  transition: 0.4s;
+  width: 90%;
+  border-radius: 10px;
+  background-color: whitesmoke;
 }
   
 .profCard {
@@ -20,7 +54,17 @@
   margin: 1rem 10% 5rem 10%;
 }
 
-/* MEDIA QUERIES aren't working yet */ 
+/* USERCARD - PROFILE */ 
+.avatar {
+  height: 150px;
+  width: 150px;
+  border-radius: 50%;
+  border: 2px solid #18a0fb;
+  box-shadow: 1px 1px 3px rgba(0, 0, 0, 0.3);
+}
+
+
+/* MEDIA QUERIES */ 
 /* Screen > Portrait Tablet */
 @media (min-width: 960px) {
     .cardContainer {
