+++ client/src/components/posts/style.css
@@ -8,11 +8,11 @@
     background-color: whitesmoke !important;
 }
 
-.profContainer {
+/* .profContainer {
     width: 100vw;
     display: flex;
     flex-direction: column;
-}
+} */
 
 .cardContainer {
     display: flex;
